LogOut added

+++ src/AstroPharm.Api/Controllers/Languages/LanguageController.cs
@@ -1,6 +1,7 @@
 ﻿using AstroPharm.Domain.Entities;
 using AstroPharm.Service.DTOs.Languages;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace AstroPharm.API.Controllers
 {
@@ -31,6 +32,14 @@ namespace AstroPharm.API.Controllers
         [HttpPost]
         public async Task<ActionResult<Language>> CreateLanguage([FromBody] LanguageForCreationDto dto)
         {
+            var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (userRole == null || (userRole != "Admin" && userRole != "SuperAdmin"))
+            {
+
+                return Unauthorized(new { message = $"{userRole} ,You are not allowed to use this method!" });
+            }
+
             var createdLanguage = await _languageService.AddAsync(dto);
             return CreatedAtAction(nameof(GetLanguage), new { id = createdLanguage.Id }, createdLanguage);
         }
@@ -38,6 +47,14 @@ namespace AstroPharm.API.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateLanguage(long id, [FromBody] LanguageForUpdateDto dto)
         {
+            var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (userRole == null || (userRole != "Admin" && userRole != "SuperAdmin"))
+            {
+
+                return Unauthorized(new { message = $"{userRole} ,You are not allowed to use this method!" });
+            }
+
             var updatedLanguage = await _languageService.ModifyAsync(id, dto);
             return Ok(updatedLanguage);
         }
@@ -45,6 +62,14 @@ namespace AstroPharm.API.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteLanguage(long id)
         {
+            var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (userRole == null || (userRole != "Admin" && userRole != "SuperAdmin"))
+            {
+
+                return Unauthorized(new { message = $"{userRole} ,You are not allowed to use this method!" });
+            }
+
             var result = await _languageService.DeleteAsync(id);
             return result ? NoContent() : NotFound();
         }
